requirements.txt
django
djangorestframework
drf-yasg
django-filter
mysqlclient
pillow
python-decouple
pyjwt
bcrypt
django-redis
